refactor(browser): split media into media and media_import handlers
